requirements.txt
streamlit
arxiv
pypdf
langchain
python-dotenv
langchain_groq
qdrant-client
langchain_community
langchain-cohere
langchain-qdrant

-e . 